feat: migrate from chalk to picocolors
This migrates away from chalk to picocolors instead, primarily because
the latter is _much_ lighter.

Keep in mind, `node-logger` still uses chalk as we use `hex(code)` which
doesn't have an alternative in picocolors. This means it will still be
in the dependency tree for many packages (for now).

This seems preferable over upgrading chalk since the latest version of
chalk will still be heavier, and is ESM-only (not sure if we're able to
adopt ESM only deps in SB yet).